Tiny improvement. See Gabber for recent changes.
Gabber is here, https://github.com/tommyettinger/Gabber . It is not a
single-file library, and it does have a dependency (RegExodus, which
allows regular expressions that work on GWT and also adds some special
custom replacement features to the regex support). This does allow it to
be quite a bit more powerful, though, and it can cipher English text
into a fictional language, randomly generate a fictional language from
any seed, and do a bit of useful processing for normal English text as
well. The README should be updated soon.